requirements.txt
python-dotenv
google-cloud-bigquery
requests
google-adk